Refactored modules around connect.js. main.js has no exports
