+++ public/index.html
@@ -70,8 +70,8 @@
 		</videos>
 
 		<buttons id="buttons" hidden>
-			<input type="button" id="changeMode" onclick="open3D()" value="Open Chat" hidden />
-			<input type="button" id="shareButton" onclick="shareScreen()" value="Share Screen" hidden />
+			<input type="button" id="changeMode" onclick="open3D()" value="Open Chat" />
+			<input type="button" id="shareButton" onclick="shareScreen()" value="Share Screen" />
 			<input type="button" id="cameraButton" onclick="shareCamera()" value="Add video" />
 		</buttons>
